+++ edge-function/component.go
@@ -1,71 +1,12 @@
 package main
 
 import (
-	"encoding/json"
 	incominghandler "example-go-component/internal/wasi/http/incoming-handler"
 	wasihttp "example-go-component/internal/wasi/http/types"
-	"example-go-component/internal/wasi/io/streams"
 	"fmt"
 
 	"go.bytecodealliance.org/cm"
 )
-
-func get_headers(request incominghandler.IncomingRequest) map[string][]string {
-	headerMap := make(map[string][]string)
-
-	headers := request.Headers()
-	entries := headers.Entries()
-	slice := entries.Slice()
-
-	for _, entry := range slice {
-		key := string(entry.F0)
-		value := string(entry.F1.Slice())
-		if _, exists := headerMap[key]; !exists {
-			headerMap[key] = []string{}
-		}
-		headerMap[key] = append(headerMap[key], value)
-	}
-
-	return headerMap
-}
-
-type Settings struct {
-	Example string `json:"example"`
-}
-
-func get_settings(request incominghandler.IncomingRequest) *Settings {
-	headerMap := get_headers(request)
-	settingsHeader := headerMap["x-edgee-component-settings"][0]
-
-	var settings Settings
-
-	err := json.Unmarshal([]byte(settingsHeader), &settings)
-	if err != nil {
-		fmt.Println("Error:", err)
-	}
-
-	return &settings
-}
-
-func get_body(request incominghandler.IncomingRequest) string {
-	body, _, _ := request.Consume().Result()
-	stream, _, _ := body.Stream().Result()
-
-	output := ""
-
-	for {
-		data, err, isErr := stream.Read(1024).Result()
-		if isErr {
-			if err == streams.StreamErrorClosed() {
-				break
-			}
-		} else {
-			output += string(data.Slice())
-		}
-	}
-
-	return output
-}
 
 func Handle(request incominghandler.IncomingRequest, responseOut incominghandler.ResponseOutparam) {
 	index := `
